requirements.txt
alembic==1.11.1
blinker==1.6.2
certifi==2023.5.7
chardet==4.0.0
click==8.1.5
colorama==0.4.6
Flask==2.3.2
Flask-Cors==4.0.0
flask-marshmallow==0.14.0
Flask-Migrate==2.0.0
Flask-Script==2.0.5
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
idna==2.10
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.19.0
packaging==23.1
PyJWT==2.7.0
PyMySQL==0.10.1
pytz==2021.1
requests==2.25.1
six==1.16.0
SQLAlchemy==2.0.18
typing_extensions==4.7.1
urllib3==1.26.16
Werkzeug==2.3.6